+++ packages/node-firebird-driver/src/lib/impl/resultset.ts
@@ -32,13 +32,13 @@ export abstract class AbstractResultSet implements ResultSet {
 	}
 
 	/**
-	 * Fetchs data from this result set.
+	 * Fetchs data from this result set as [col1, col2, ..., colN][].
 	 *
 	 * If an exception is found after fetching a row but before reaching options.fetchSize, it's throw is delayed for the next fetch call.
 	 *
 	 * If result set has no more rows, returns an empty array.
 	 */
-	async fetch(options?: FetchOptions): Promise<Array<Array<any>>> {
+	async fetch(options?: FetchOptions): Promise<any[][]> {
 		this.check();
 
 		if (this.finished)
@@ -54,11 +54,37 @@ export abstract class AbstractResultSet implements ResultSet {
 		return fetchRet.rows;
 	}
 
+	/**
+	 * Fetchs data from this result set as T[].
+	 * Where <T> represents your object interface.
+	 *
+	 * If an exception is found after fetching a row but before reaching options.fetchSize, it's throw is delayed for the next fetch call.
+	 *
+	 * If result set has no more rows, returns an empty array.
+	 */
+	async fetchAsObject<T extends object>(options?: FetchOptions): Promise<T[]> {
+		const array = await this.fetch(options);
+		const cols = (await this.statement?.columnLabels) || [];
+
+		return array.map(row => {
+			const obj = {} as T;
+
+			// Loop on row column value.
+			row.forEach((v: any, idx: number) => {
+				const col = cols[idx];
+				(obj as any)[col] = v;
+			});
+
+			return obj;
+		});
+	}
+
 	private check() {
 		if (!this.statement)
 			throw new Error('ResultSet is already closed.');
 	}
 
 	protected abstract async internalClose(): Promise<void>;
-	protected abstract async internalFetch(options?: FetchOptions): Promise<{ finished: boolean; rows: Array<Array<any>> }>;
+
+	protected abstract async internalFetch(options?: FetchOptions): Promise<{ finished: boolean; rows: any[][] }>;
 }